infra_cdk/requirements.txt
aws-cdk-lib==2.195.0
constructs==10.4.2
python-dotenv==1.0.1
